Move dto classes from util to 'dto' package
