+++ src/main/java/com/a_smart_cookie/controller/command/admin/UsersManagementCommand.java
@@ -8,7 +8,7 @@ import com.a_smart_cookie.dto.admin.UserForManagement;
 import com.a_smart_cookie.exception.ServiceException;
 import com.a_smart_cookie.service.ServiceFactory;
 import com.a_smart_cookie.service.UserService;
-import com.a_smart_cookie.util.pagination.PaginationHandler;
+import com.a_smart_cookie.util.PaginationHandler;
 import org.apache.log4j.Logger;
 
 import javax.servlet.http.HttpServletRequest;
