+++ src/main/java/com/a_smart_cookie/dto/catalog/SortingDirection.java
@@ -1,4 +1,4 @@
-package com.a_smart_cookie.util.sorting;
+package com.a_smart_cookie.dto.catalog;
 
 import com.a_smart_cookie.entity.Language;
 import com.a_smart_cookie.util.StringHandler;
