+++ src/main/webapp/WEB-INF/jsp/catalog.jsp
@@ -1,7 +1,7 @@
 <%@ page pageEncoding="UTF-8" %>
-<%@ page import="com.a_smart_cookie.util.sorting.SortingParameter" %>
-<%@ page import="com.a_smart_cookie.util.sorting.SortingDirection" %>
-<%@ page import="com.a_smart_cookie.util.pagination.ItemsPerPage" %>
+<%@ page import="com.a_smart_cookie.dto.catalog.SortingParameter" %>
+<%@ page import="com.a_smart_cookie.dto.catalog.SortingDirection" %>
+<%@ page import="com.a_smart_cookie.dto.ItemsPerPage" %>
 <%@ include file="/WEB-INF/jspf/directive/page.jspf" %>
 <%@ include file="/WEB-INF/jspf/directive/taglib.jspf" %>
 <%@ include file="/WEB-INF/jspf/directive/locale.jspf" %>
